Dev nch

+++ src/backend/rest/accounts.py
@@ -18,9 +18,10 @@
 import os
 import base64
 import mimetypes
+from flask_babel import gettext
 from src.backend.functions import retrieve_custom_from_url
 from src.backend.main import create_classes_from_custom_id
-from src.backend.import_controllers import auth, accounts, verifier
+from src.backend.import_controllers import auth, accounts, verifier, privileges
 from flask import Blueprint, request, make_response, jsonify, g as current_context
 
 bp = Blueprint('accounts', __name__, url_prefix='/ws/')
@@ -29,31 +30,20 @@ bp = Blueprint('accounts', __name__, url_prefix='/ws/')
 @bp.route('accounts/suppliers/list', methods=['GET'])
 @auth.token_required
 def suppliers_list():
-    args = {
-        'select': ['*', 'count(*) OVER() as total'],
-        'where': ['status <> %s'],
-        'data': ['DEL'],
-        'offset': request.args['offset'] if 'offset' in request.args else 0,
-        'limit': request.args['limit'] if 'limit' in request.args else 'ALL',
-        'order_by': [request.args['order']] if 'order' in request.args else ''
-    }
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/accounts/suppliers/list'}), 403
 
-    if 'search' in request.args and request.args['search']:
-        args['offset'] = ''
-        args['where'].append(
-            "(LOWER(name) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(siret) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(email) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(siren) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(vat_number) LIKE '%%" + request.args['search'].lower() + "%%')"
-        )
-    res = accounts.retrieve_suppliers(args)
+    res = accounts.retrieve_suppliers(request.args)
     return make_response(res[0], res[1])
 
 
 @bp.route('accounts/suppliers/getById/<int:supplier_id>', methods=['GET'])
 @auth.token_required
 def get_supplier_by_id(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/getById/{supplier_id}'}), 403
+
     supplier = accounts.get_supplier_by_id(supplier_id)
     return make_response(jsonify(supplier[0])), supplier[1]
 
@@ -61,6 +51,11 @@ def get_supplier_by_id(supplier_id):
 @bp.route('accounts/getAdressById/<int:address_id>', methods=['GET'])
 @auth.token_required
 def get_address_by_id(address_id):
+    if not privileges.has_privileges(request.environ['user_id'],
+                                     ['update_supplier | update_customer | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/getById/{address_id}'}), 403
+
     _address = accounts.get_address_by_id(address_id)
     return make_response(jsonify(_address[0])), _address[1]
 
@@ -68,36 +63,22 @@ def get_address_by_id(address_id):
 @bp.route('accounts/suppliers/update/<int:supplier_id>', methods=['PUT'])
 @auth.token_required
 def update_supplier(supplier_id):
-    data = request.json['args']
-    _set = {}
-    if 'address_id' in data:
-        _set.update({'address_id': data['address_id']})
-    if 'name' in data:
-        _set.update({'name': data['name']})
-    if 'siret' in data:
-        _set.update({'siret': data['siret']})
-    if 'siren' in data:
-        _set.update({'siren': data['siren']})
-    if 'iban' in data:
-        _set.update({'iban': data['iban']})
-    if 'email' in data:
-        _set.update({'email': data['email']})
-    if 'vat_number' in data:
-        _set.update({'vat_number': data['vat_number']})
-    if 'form_id' in data:
-        _set.update({'form_id': data['form_id']})
-    if 'get_only_raw_footer' in data:
-        _set.update({'get_only_raw_footer': data['get_only_raw_footer']})
-    if 'document_lang' in data:
-        _set.update({'document_lang': data['document_lang']})
+    if not privileges.has_privileges(request.environ['user_id'], ['update_supplier | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/update/{supplier_id}'}), 403
 
-    res = accounts.update_supplier(supplier_id, _set)
+    data = request.json['args']
+    res = accounts.update_supplier(supplier_id, data)
     return make_response(jsonify(res[0])), res[1]
 
 
 @bp.route('accounts/supplier/<int:supplier_id>/updatePosition', methods=['PUT'])
 @auth.token_required
 def update_position(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['update_supplier | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/{supplier_id}/updatePosition'}), 403
+
     data = request.json['args']
     res = accounts.update_position_by_supplier_id(supplier_id, data)
     return make_response(res[0], res[1])
@@ -106,6 +87,10 @@ def update_position(supplier_id):
 @bp.route('accounts/supplier/<int:supplier_id>/updatePage', methods=['PUT'])
 @auth.token_required
 def update_page(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['update_supplier | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/{supplier_id}/updatePage'}), 403
+
     data = request.json['args']
     res = accounts.update_page_by_supplier_id(supplier_id, data)
     return make_response(res[0], res[1])
@@ -114,6 +99,11 @@ def update_page(supplier_id):
 @bp.route('accounts/addresses/update/<int:address_id>', methods=['PUT'])
 @auth.token_required
 def update_address(address_id):
+    if not privileges.has_privileges(request.environ['user_id'],
+                                     ['update_supplier | update_customer | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/addresses/update/{address_id}'}), 403
+
     data = request.json['args']
     res = accounts.update_address(address_id, data)
     return make_response(jsonify(res[0])), res[1]
@@ -122,6 +112,10 @@ def update_address(address_id):
 @bp.route('accounts/addresses/updateBySupplierId/<int:suplier_id>', methods=['PUT'])
 @auth.token_required
 def update_address_by_supplier_id(suplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['update_supplier | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/addresses/updateBySupplierId/{suplier_id}'}), 403
+
     data = request.json['args']
     res = accounts.update_address_by_supplier_id(suplier_id, data)
     return make_response(jsonify(res[0])), res[1]
@@ -130,6 +124,10 @@ def update_address_by_supplier_id(suplier_id):
 @bp.route('accounts/addresses/create', methods=['POST'])
 @auth.token_required
 def create_address():
+    if not privileges.has_privileges(request.environ['user_id'],
+                                     ['create_supplier | create_customer | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/accounts/addresses/create'}), 403
+
     data = request.json['args']
     module = ''
     if 'module' in data:
@@ -137,6 +135,7 @@ def create_address():
         del data['module']
     if all(data[x] is None for x in data) and module == 'splitter':
         return make_response(''), 204
+
     res = accounts.create_address(data)
     return make_response(jsonify(res[0])), res[1]
 
@@ -144,6 +143,9 @@ def create_address():
 @bp.route('accounts/suppliers/create', methods=['POST'])
 @auth.token_required
 def create_supplier():
+    if not privileges.has_privileges(request.environ['user_id'], ['create_supplier | access_verifier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/accounts/suppliers/create'}), 403
+
     data = request.json['args']
     res = accounts.create_supplier(data)
     return make_response(jsonify(res[0])), res[1]
@@ -152,6 +154,10 @@ def create_supplier():
 @bp.route('accounts/suppliers/delete/<int:supplier_id>', methods=['DELETE'])
 @auth.token_required
 def delete_supplier(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list', 'update_supplier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/delete/{supplier_id}'}), 403
+
     res = accounts.delete_supplier(supplier_id)
     return make_response(jsonify(res[0])), res[1]
 
@@ -159,6 +165,10 @@ def delete_supplier(supplier_id):
 @bp.route('accounts/suppliers/deletePositions/<int:supplier_id>', methods=['DELETE'])
 @auth.token_required
 def delete_supplier_positions(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list', 'update_supplier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/deletePositions/{supplier_id}'}), 403
+
     res = accounts.update_supplier(supplier_id, {'positions': '{}'})
     return make_response(jsonify(res[0])), res[1]
 
@@ -166,6 +176,10 @@ def delete_supplier_positions(supplier_id):
 @bp.route('accounts/suppliers/<int:supplier_id>/deletePosition', methods=['PUT'])
 @auth.token_required
 def delete_supplier_position(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list', 'update_supplier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/{supplier_id}/deletePosition'}), 403
+
     form_id = request.json['args']['form_id']
     args = request.json['args']
     res = '', 200
@@ -182,6 +196,10 @@ def delete_supplier_position(supplier_id):
 @bp.route('accounts/suppliers/<int:supplier_id>/deletePage', methods=['PUT'])
 @auth.token_required
 def delete_supplier_page(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list', 'update_supplier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/{supplier_id}/deletePage'}), 403
+
     form_id = request.json['args']['form_id']
     args = request.json['args']
     res = '', 200
@@ -198,6 +216,10 @@ def delete_supplier_page(supplier_id):
 @bp.route('accounts/suppliers/skipAutoValidate/<int:supplier_id>', methods=['PUT'])
 @auth.token_required
 def skip_auto_validate(supplier_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list', 'update_supplier']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/suppliers/skipAutoValidate/{supplier_id}'}), 403
+
     res = accounts.update_supplier(supplier_id, {'skip_auto_validate': True})
     return make_response(jsonify(res[0])), res[1]
 
@@ -206,30 +228,20 @@ def skip_auto_validate(supplier_id):
 @bp.route('accounts/customers/list/<string:module>', methods=['GET'])
 @auth.token_required
 def customers_list(module=False):
-    args = {
-        'select': ['*', 'count(*) OVER() as total'],
-        'where': ['status <> %s', 'module = %s'] if module else ['status <> %s'],
-        'data': ['DEL', module] if module else ['DEL'],
-        'offset': request.args['offset'] if 'offset' in request.args else 0,
-        'limit': request.args['limit'] if 'limit' in request.args else 'ALL'
-    }
-    if 'search' in request.args and request.args['search']:
-        args['offset'] = ''
-        args['where'].append(
-            "(LOWER(name) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(siret) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(company_number) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(siren) LIKE '%%" + request.args['search'].lower() + "%%' OR "
-            "LOWER(vat_number) LIKE '%%" + request.args['search'].lower() + "%%')"
-        )
+    if not privileges.has_privileges(request.environ['user_id'], ['customers_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/accounts/customers/list'}), 403
 
-    res = accounts.retrieve_customers(args)
+    res = accounts.retrieve_customers(request.args, module)
     return make_response(res[0], res[1])
 
 
 @bp.route('accounts/customers/getById/<int:customer_id>', methods=['GET'])
 @auth.token_required
 def get_customer_by_id(customer_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['customers_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/customers/getById/{customer_id}'}), 403
+
     _customer = accounts.get_customer_by_id(customer_id)
     return make_response(jsonify(_customer[0])), _customer[1]
 
@@ -237,6 +249,10 @@ def get_customer_by_id(customer_id):
 @bp.route('accounts/customers/update/<int:customer_id>', methods=['PUT'])
 @auth.token_required
 def update_customer(customer_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['update_customer']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/customers/update/{customer_id}'}), 403
+
     data = request.json['args']
     res = accounts.update_customer(customer_id, data)
     return make_response(jsonify(res[0])), res[1]
@@ -245,6 +261,9 @@ def update_customer(customer_id):
 @bp.route('accounts/customers/create', methods=['POST'])
 @auth.token_required
 def create_customer():
+    if not privileges.has_privileges(request.environ['user_id'], ['create_customer']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/accounts/customers/create'}), 403
+
     data = request.json['args']
     res = accounts.create_customer(data)
     return make_response(jsonify(res[0])), res[1]
@@ -253,6 +272,10 @@ def create_customer():
 @bp.route('accounts/customers/delete/<int:customer_id>', methods=['DELETE'])
 @auth.token_required
 def delete_customer(customer_id):
+    if not privileges.has_privileges(request.environ['user_id'], ['customers_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'),
+                        'message': f'/accounts/customers/delete/{customer_id}'}), 403
+
     res = accounts.delete_customer(customer_id)
     return make_response(jsonify(res[0])), res[1]
 
@@ -286,12 +309,16 @@ def get_reference_file():
     file_path = docservers['REFERENTIALS_PATH'] + '/' + config['REFERENCIAL']['referencialsupplierdocument']
     mime = mimetypes.guess_type(file_path)[0]
     file_content = verifier.get_file_content('referential_supplier', os.path.basename(file_path), mime)
-    return make_response({'filename': os.path.basename(file_path), 'mimetype': mime, 'file': str(base64.b64encode(file_content.get_data()).decode('UTF-8'))}), 200
+    return make_response({'filename': os.path.basename(file_path), 'mimetype': mime,
+                          'file': str(base64.b64encode(file_content.get_data()).decode('UTF-8'))}), 200
 
 
 @bp.route('accounts/supplier/importSuppliers', methods=['POST'])
 @auth.token_required
 def import_suppliers():
+    if not privileges.has_privileges(request.environ['user_id'], ['suppliers_list']):
+        return jsonify({'errors': gettext('UNAUTHORIZED_ROUTE'), 'message': '/accounts/supplier/importSuppliers'}), 403
+
     files = request.files
     res = '', 200
     if files:
